+++ src/modules/active-player-switch/APSModule.cpp
@@ -1,5 +1,6 @@
 #include "APSModule.hpp"
 #include "Hooks.hpp"
+#include <globed/core/RoomManager.hpp>
 #include <core/hooks/GJBaseGameLayer.hpp>
 
 using namespace geode::prelude;
@@ -10,6 +11,12 @@ APSModule::APSModule() {}
 
 void APSModule::onModuleInit() {
     this->setAutoEnableMode(AutoEnableMode::Level);
+}
+
+void APSModule::onJoinLevel(GlobedGJBGL* gjbgl, GJGameLevel* level, bool editor) {
+    if (!RoomManager::get().getSettings().switcheroo) {
+        (void) this->disable();
+    }
 }
 
 void APSModule::onPlayerDeath(GlobedGJBGL* gjbgl, RemotePlayer* player, const PlayerDeath& death) {