+++ src/modules/active-player-switch/APSModule.hpp
@@ -26,6 +26,7 @@ public:
         return "";
     }
 private:
+    void onJoinLevel(GlobedGJBGL* gjbgl, GJGameLevel* level, bool editor) override;
     void onPlayerDeath(GlobedGJBGL* gjbgl, RemotePlayer* player, const PlayerDeath& death) override;
 };