+++ src/ui/game/VoiceOverlay.hpp
@@ -16,6 +16,7 @@ public:
 
 private:
     std::unordered_map<int, VoiceOverlayCell*> m_cells;
+    float m_threshold = 0.f;
 
     bool init() override;