+++ src/ui/game/VoiceOverlayCell.hpp
@@ -3,6 +3,7 @@
 #include <globed/prelude.hpp>
 #include <globed/core/data/PlayerDisplayData.hpp>
 #include <ui/misc/AudioVisualizer.hpp>
+#include <asp/time/Instant.hpp>
 
 namespace globed {
 
@@ -11,6 +12,7 @@ public:
     static VoiceOverlayCell* create(const PlayerDisplayData& data);
 
     void updateLoudness(float loudness);
+    asp::time::Duration sinceLastSpoken();
 
     inline int getAccountId() {
         return m_accountId;
@@ -19,6 +21,7 @@ public:
 private:
     AudioVisualizer* m_visualizer;
     CCNode* m_wrapper;
+    asp::time::Instant m_lastSpoken = asp::time::Instant::now();
     int m_accountId;
 
     bool init(const PlayerDisplayData& data);